move error and json parsing functionality outside dispatch
