+++ packages/router/src/errors.ts
@@ -7,8 +7,9 @@
 
 import {randomUUID} from 'crypto';
 import {getRouterOptions} from './router';
-import {statusCodeToReasonPhrase} from './status-codes';
-import {Obj, RouteErrorParams} from './types';
+import {StatusCodes, statusCodeToReasonPhrase} from './status-codes';
+import {Obj, PublicError, Response, RouteErrorParams} from './types';
+import {stringifyResponseBody} from './jsonBodyParser';
 
 export class RouteError extends Error {
     /** id of the error, if RouterOptions.autoGenerateErrorId is set to true and id with timestamp+uuid will be generated */
@@ -31,4 +32,49 @@ export class RouteError extends Error {
         this.publicData = publicData;
         Object.setPrototypeOf(this, RouteError.prototype);
     }
+}
+
+/**
+ * This is a function to be called from outside the router.
+ * Whenever there is an error outside the router, this function should be called.
+ * So error keep the same format as when they were generated inside the router.
+ * This also stringifies public errors into response.json.
+ * @param routeResponse
+ * @param originalError
+ */
+
+export function getResponseFromError(
+    originalError: any,
+    statusCode: StatusCodes = StatusCodes.INTERNAL_SERVER_ERROR,
+    publicMessage = 'Internal Error'
+): Response {
+    const routerOptions = getRouterOptions();
+    const error = new RouteError({
+        statusCode,
+        publicMessage,
+        originalError,
+    });
+    const publicErrors = [getPublicErrorFromRouteError(error)];
+    const response = {
+        statusCode,
+        publicErrors,
+        headers: {},
+        body: {},
+        json: '',
+    } as Response;
+
+    stringifyResponseBody(response, routerOptions);
+    return response;
+}
+
+export function getPublicErrorFromRouteError(routeError: RouteError): PublicError {
+    // creating a new public error object to avoid exposing the original error
+    const publicError: PublicError = {
+        name: routeError.name,
+        statusCode: routeError.statusCode,
+        message: routeError.publicMessage,
+    };
+    if (routeError.id) publicError.id = routeError.id;
+    if (routeError.publicData) publicError.errorData = routeError.publicData;
+    return publicError;
 }
